Implemented Yihua's comments, round 3.
