+++ Makefile
@@ -1278,7 +1278,6 @@ ref_counted_ptr_test: $(BINDIR)/$(CONFIG)/ref_counted_ptr_test
 ref_counted_test: $(BINDIR)/$(CONFIG)/ref_counted_test
 retry_throttle_test: $(BINDIR)/$(CONFIG)/retry_throttle_test
 s2a_credentials_test: $(BINDIR)/$(CONFIG)/s2a_credentials_test
-s2a_handshaker_client_caller_test: $(BINDIR)/$(CONFIG)/s2a_handshaker_client_caller_test
 s2a_handshaker_client_test: $(BINDIR)/$(CONFIG)/s2a_handshaker_client_test
 s2a_record_protocol_test: $(BINDIR)/$(CONFIG)/s2a_record_protocol_test
 s2a_security_connector_test: $(BINDIR)/$(CONFIG)/s2a_security_connector_test
@@ -1754,7 +1753,6 @@ buildtests_cxx: privatelibs_cxx \
   $(BINDIR)/$(CONFIG)/ref_counted_test \
   $(BINDIR)/$(CONFIG)/retry_throttle_test \
   $(BINDIR)/$(CONFIG)/s2a_credentials_test \
-  $(BINDIR)/$(CONFIG)/s2a_handshaker_client_caller_test \
   $(BINDIR)/$(CONFIG)/s2a_handshaker_client_test \
   $(BINDIR)/$(CONFIG)/s2a_record_protocol_test \
   $(BINDIR)/$(CONFIG)/s2a_security_connector_test \
@@ -1932,7 +1930,6 @@ buildtests_cxx: privatelibs_cxx \
   $(BINDIR)/$(CONFIG)/ref_counted_test \
   $(BINDIR)/$(CONFIG)/retry_throttle_test \
   $(BINDIR)/$(CONFIG)/s2a_credentials_test \
-  $(BINDIR)/$(CONFIG)/s2a_handshaker_client_caller_test \
   $(BINDIR)/$(CONFIG)/s2a_handshaker_client_test \
   $(BINDIR)/$(CONFIG)/s2a_record_protocol_test \
   $(BINDIR)/$(CONFIG)/s2a_security_connector_test \
@@ -2460,8 +2457,6 @@ test_cxx: buildtests_cxx
 	$(Q) $(BINDIR)/$(CONFIG)/retry_throttle_test || ( echo test retry_throttle_test failed ; exit 1 )
 	$(E) "[RUN]     Testing s2a_credentials_test"
 	$(Q) $(BINDIR)/$(CONFIG)/s2a_credentials_test || ( echo test s2a_credentials_test failed ; exit 1 )
-	$(E) "[RUN]     Testing s2a_handshaker_client_caller_test"
-	$(Q) $(BINDIR)/$(CONFIG)/s2a_handshaker_client_caller_test || ( echo test s2a_handshaker_client_caller_test failed ; exit 1 )
 	$(E) "[RUN]     Testing s2a_handshaker_client_test"
 	$(Q) $(BINDIR)/$(CONFIG)/s2a_handshaker_client_test || ( echo test s2a_handshaker_client_test failed ; exit 1 )
 	$(E) "[RUN]     Testing s2a_record_protocol_test"
@@ -19229,49 +19224,6 @@ deps_s2a_credentials_test: $(S2A_CREDENTIALS_TEST_OBJS:.o=.dep)
 ifneq ($(NO_SECURE),true)
 ifneq ($(NO_DEPS),true)
 -include $(S2A_CREDENTIALS_TEST_OBJS:.o=.dep)
-endif
-endif
-
-
-S2A_HANDSHAKER_CLIENT_CALLER_TEST_SRC = \
-    test/core/tsi/s2a/s2a_handshaker_client_caller_test.cc \
-
-S2A_HANDSHAKER_CLIENT_CALLER_TEST_OBJS = $(addprefix $(OBJDIR)/$(CONFIG)/, $(addsuffix .o, $(basename $(S2A_HANDSHAKER_CLIENT_CALLER_TEST_SRC))))
-ifeq ($(NO_SECURE),true)
-
-# You can't build secure targets if you don't have OpenSSL.
-
-$(BINDIR)/$(CONFIG)/s2a_handshaker_client_caller_test: openssl_dep_error
-
-else
-
-
-
-
-ifeq ($(NO_PROTOBUF),true)
-
-# You can't build the protoc plugins or protobuf-enabled targets if you don't have protobuf 3.5.0+.
-
-$(BINDIR)/$(CONFIG)/s2a_handshaker_client_caller_test: protobuf_dep_error
-
-else
-
-$(BINDIR)/$(CONFIG)/s2a_handshaker_client_caller_test: $(PROTOBUF_DEP) $(S2A_HANDSHAKER_CLIENT_CALLER_TEST_OBJS) $(LIBDIR)/$(CONFIG)/libgrpc_test_util.a $(LIBDIR)/$(CONFIG)/libgrpc.a
-	$(E) "[LD]      Linking $@"
-	$(Q) mkdir -p `dirname $@`
-	$(Q) $(LDXX) $(LDFLAGS) $(S2A_HANDSHAKER_CLIENT_CALLER_TEST_OBJS) $(LIBDIR)/$(CONFIG)/libgrpc_test_util.a $(LIBDIR)/$(CONFIG)/libgrpc.a $(LDLIBSXX) $(LDLIBS_PROTOBUF) $(LDLIBS) $(LDLIBS_SECURE) $(GTEST_LIB) -o $(BINDIR)/$(CONFIG)/s2a_handshaker_client_caller_test
-
-endif
-
-endif
-
-$(OBJDIR)/$(CONFIG)/test/core/tsi/s2a/s2a_handshaker_client_caller_test.o:  $(LIBDIR)/$(CONFIG)/libgrpc_test_util.a $(LIBDIR)/$(CONFIG)/libgrpc.a
-
-deps_s2a_handshaker_client_caller_test: $(S2A_HANDSHAKER_CLIENT_CALLER_TEST_OBJS:.o=.dep)
-
-ifneq ($(NO_SECURE),true)
-ifneq ($(NO_DEPS),true)
--include $(S2A_HANDSHAKER_CLIENT_CALLER_TEST_OBJS:.o=.dep)
 endif
 endif
